Changes so multiple blocks of same paradigm type can be analysed at once
I've made it so there are 10 reps of each condition in each paradigm,
which means we'll have three repeats of each paradigm interleaved. So
that we can still analyse all of the data from each separate block of
the same paradigm, I've altered the basic analysis file so that it will
take all of the data files of a certain condition and subject (currently
still have to enter manually but want to automate) and combine them so
they can be analysed together.

Also added some comments.