Improvements for psychometric analysis
Now automatically saves the raw data from that participant and
condition as an excel readable (.csv) file.

Now saves the key psychometric data from all conditions for that
participant into a .csv file so that these values can be compared
later.

Edits so that if the condition is lateral, the x axis label on the
graph that is generated refers to cm/s moved on the screen, and if it
is a depth condition, the x label refers to cm/s in depth.

Improvements to how the figure is automatically saved.