Playing around with psychometric analysis files
fixed some bugs — all of the i = 1:whatever are now iCond or IAnalysis
so that they’re not overlapping. lambda is now 0 as that seems to be
where most of the major convergence failures were coming from. the
confidence interval and SE stuff for the threshold and slope values
that i’m finding is not yet sorted.