Beginnings of a file to generate summary graphs from the psychometric analysis
hasn’t yet been tested on the lab computer, so will need to be
debugged. Should create plots of the thresholds (with confidence
interval error bars) and slopes (again with CI error bars) for each
participant with every analysis type and save the graphs that are
generated. Will also need to get it so that the graphs that are
produced look aesthetically pleasing.