Start of updates to psychometric analysis to allow use with experiment 2 data
Started to update the psychometric analysis file so that it will run
with the data from experiment 2 as well as that from experiment 1.

I’ve added a few comments for things that I can’t change just yet, e.g
the condition list, file location and so on. I’ve created an additional
analysis type so that the results of experiment 2 can be analysed in
terms of the world speed. I’ve also started adding looming numbers to
all of the analysis types, but this still needs to be done for full and
full_arcmin. Also took out the stuff for start_arcmin and end_arcmin
because I don’t think I need it.